requirements.txt
langchain==0.0.300
mypy-extensions==1.0.0
numexpr==2.8.4
python-environ==0.4.54
SQLAlchemy==2.0.11
qdrant_client==0.11.9
pyodbc==4.0.39
streamlit==1.27.0
python-dotenv==1.0.0
pymongo==4.5.0
# langchain_experimenta==0.0.20